backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
PyJWT==2.8.0
python-decouple==3.8
Werkzeug==3.0.1
PyMySQL==1.1.0 